fix type cast

+++ android/src/main/java/it/iotinga/blelibrary/RNBleManager.kt
@@ -177,15 +177,15 @@ class RNBleManager(
 
     if (options?.contains("timeout") == true) {
       log(Log.DEBUG, "set timeout to ${options["timeout"]}")
-      request.timeout(options["timeout"] as Long)
+      request.timeout((options["timeout"] as Double).toLong())
     }
     if (options?.contains("preferredPhy") == true) {
       log(Log.DEBUG, "set preferredPhy to ${options["preferredPhy"]}")
-      request.usePreferredPhy(options["preferredPhy"] as Int)
+      request.usePreferredPhy((options["preferredPhy"] as Double).toInt())
     }
     if (options?.contains("logLevel") == true) {
       log(Log.DEBUG, "set logLevel to ${options["logLevel"]}")
-      logLevel = options["logLevel"] as Int
+      logLevel = (options["logLevel"] as Double).toInt()
     }
 
     request.enqueue()
